Fixed: compatibility support with wordpress 6.5 version
