+++ block.json
@@ -17,6 +17,5 @@
 		"plan"
 	],
 	"description": "EB Pricing Table will let you create effective product pricing table with perfect styling to get more sales from your prospective buyers.",
-	"textdomain": "price-table-block",
-	"editorScript": "file:./dist/index.js"
+	"textdomain": "price-table-block"
 }
