+++ config/entries.js
@@ -1,5 +1,5 @@
 //Export All Controls
-import "../controls/src/backend-css";
+import "../controls/src/backend.scss";
 
 //Export All Controls
 export { default as ResponsiveDimensionsControl } from "../controls/src/controls/dimensions-control-v2";
@@ -9,6 +9,9 @@ export { default as BorderShadowControl } from "../controls/src/controls/border-
 export { default as BackgroundControl } from "../controls/src/controls/background-control";
 export { default as ResponsiveRangeController } from "../controls/src/controls/responsive-range-control";
 export { default as faIcons } from "../controls/src/extras/faIcons";
+export { default as DynamicInputValueHandler } from "../controls/src/controls/dynamic-field/DynamicInputValueHandler";
+export { EBIconPicker, EBDisplayIcon } from "../controls/src/controls/icon-picker";
+export { default as DynamicInputControl } from "../controls/src/controls/dynamic-field";
 
 import "../controls/src/group-controls";
 export { default as AdvancedControls } from "../controls/src/group-controls/components/advanced-controls";
@@ -28,4 +31,5 @@ export {
 	generateBorderShadowAttributes,
 	generateResponsiveRangeAttributes,
 	ebConditionalRegisterBlockType,
+	StyleComponent
 } from "../controls/src/helpers";
