+++ price-table-block.php
@@ -4,7 +4,7 @@
  * Plugin Name:     Price Table Block
  * Plugin URI:         https://essential-blocks.com
  * Description:     Instantly create beautiful pricing menu for eCommerce website
- * Version:         1.2.5
+ * Version:         1.2.6
  * Author:          WPDeveloper
  * Author URI:         https://wpdeveloper.net
  * License:         GPL-3.0-or-later
@@ -28,7 +28,7 @@ require_once __DIR__ . '/lib/style-handler/style-handler.php';
 
 function create_block_pricing_table_block_init() {
 
-    define( 'PRICE_TABLE_BLOCKS_VERSION', "1.2.5" );
+    define( 'PRICE_TABLE_BLOCKS_VERSION', " 1.2.6" );
     define( 'PRICE_TABLE_BLOCKS_ADMIN_URL', plugin_dir_url( __FILE__ ) );
     define( 'PRICE_TABLE_BLOCKS_ADMIN_PATH', dirname( __FILE__ ) );
