+++ src/example.js
