+++ src/index.js
@@ -2,7 +2,6 @@
  * WordPress dependeincies
  */
 import { __ } from "@wordpress/i18n";
-import { registerBlockType } from "@wordpress/blocks";
 
 /**
  * Internal dependencies
