+++ src/style.scss
@@ -29,15 +29,18 @@
 	left: 0;
 	z-index: 1;
 }
+
 .eb-pricing-item div:not(.eb-pricing-item-overlay) {
 	position: relative;
 	z-index: 11 !important;
 }
 
 .eb-pricing .eb-pricing-button {
+	display: inline-block;
 	padding: 12px 25px;
 	background: #00c853;
 	font-size: 14px;
+	line-height: 16px;
 	font-weight: 600;
 	color: #fff;
 	text-transform: uppercase;
@@ -76,6 +79,7 @@
 	text-decoration: line-through;
 	padding-right: 8px;
 }
+
 .eb-pricing .eb-pricing-item ul li a {
 	text-decoration: none;
 }
@@ -401,6 +405,7 @@
 	z-index: 9;
 	opacity: 0.9;
 }
+
 .eb-pricing.style-2 .eb-pricing-item.ribbon-2 {
 	&.left {
 		&:before {
@@ -598,7 +603,7 @@
 
 /*--- Media Query ---*/
 
-@media only screen and (min-width: 768px) and (max-width: 992px) {
+@media (min-width: 768px) and (max-width: 992px) {
 	.eb-pricing {
 		display: block;
 	}
@@ -611,7 +616,7 @@
 	}
 }
 
-@media only screen and (max-width: 480px) {
+@media (max-width: 480px) {
 	.eb-pricing {
 		display: block;
 	}
@@ -638,10 +643,12 @@
 	padding: 30px;
 	color: #fff;
 }
+
 .eb-pricing.style-3 .eb-pricing-item.featured {
 	position: relative;
 	z-index: 0;
 }
+
 .eb-pricing.style-3 .eb-pricing-item.ribbon-1:before {
 	content: "";
 	position: absolute;
@@ -654,6 +661,7 @@
 	z-index: 1;
 	border-radius: 5px 5px 0px 0px;
 }
+
 .eb-pricing.style-3 .eb-pricing-item.ribbon-1.bottom:before {
 	content: "";
 	position: absolute;
@@ -667,6 +675,7 @@
 	z-index: 1;
 	border-radius: 0px 0px 5px 5px;
 }
+
 .eb-pricing.style-3 .eb-pricing-item.ribbon-2:before {
 	content: "Featured";
 	position: absolute;
@@ -680,6 +689,7 @@
 	text-transform: uppercase;
 	padding: 5px 10px;
 }
+
 .eb-pricing.style-3 .eb-pricing-item.ribbon-2:after {
 	content: "";
 	position: absolute;
@@ -692,6 +702,7 @@
 	z-index: 9;
 	opacity: 0.9;
 }
+
 .eb-pricing.style-3 .eb-pricing-item.ribbon-2 {
 	&.left {
 		&:before {
@@ -724,6 +735,7 @@
 		}
 	}
 }
+
 .eb-pricing.style-3 .eb-pricing-item.ribbon-3:before {
 	content: "Featured";
 	position: absolute;
@@ -737,10 +749,12 @@
 	text-transform: uppercase;
 	padding: 5px 15px;
 }
+
 .eb-pricing.style-3 .eb-pricing-item.ribbon-3.left:before {
 	left: 15px;
 	right: auto;
 }
+
 .eb-pricing.style-3 .eb-pricing-item .header,
 .eb-pricing.style-3 .eb-pricing-item .eb-pricing-header {
 	position: relative;
@@ -758,14 +772,17 @@
 	line-height: 40px;
 	margin: 0px;
 }
+
 .eb-pricing.style-3 .eb-pricing-item .header .eb-princing-subititle,
 .eb-pricing.style-3 .eb-pricing-item .eb-pricing-header .eb-princing-subititle {
 	font-size: 14px;
 	color: rgba(255, 255, 255, 0.5);
 }
+
 .eb-pricing.style-3 .eb-pricing-item .eb-pricing-tag {
 	margin-bottom: 30px;
 }
+
 .eb-pricing.style-3 .eb-pricing-item .eb-pricing-tag .price-tag {
 	position: relative;
 	display: inline-block;
@@ -775,14 +792,17 @@
 	padding-left: 12px;
 	z-index: 0;
 }
+
 .eb-pricing.style-3 .eb-pricing-item .price-tag .price-currency {
 	font-size: 36px;
 	font-weight: 700;
 }
+
 .eb-pricing.style-3 .eb-pricing-item .body ul,
 .eb-pricing.style-3 .eb-pricing-item .eb-pricing-body ul {
 	margin-bottom: 30px;
 }
+
 .eb-pricing.style-3 .eb-pricing-item .body ul li,
 .eb-pricing.style-3 .eb-pricing-item .eb-pricing-body ul li {
 	display: block;
@@ -793,17 +813,19 @@
 	color: rgba(255, 255, 255, 0.7);
 	border-bottom: 1px solid rgba(9, 9, 9, 0.04);
 }
+
 .eb-pricing.style-3 .eb-pricing-item .body ul li:last-child,
 .eb-pricing.style-3 .eb-pricing-item .eb-pricing-body ul li:last-child {
 	border: none;
 }
+
 .eb-pricing.style-3 .eb-pricing-item .eb-pricing-tag.on-top {
 	margin-top: 15px;
 	margin-bottom: 15px;
 }
 
 /*--- Media Query ---*/
-@media only screen and (min-width: 768px) and (max-width: 992px) {
+@media (min-width: 768px) and (max-width: 992px) {
 	.eb-pricing {
 		display: block;
 	}
@@ -814,7 +836,8 @@
 		margin: 0 auto 30px auto;
 	}
 }
-@media only screen and (max-width: 480px) {
+
+@media (max-width: 480px) {
 	.eb-pricing {
 		display: block;
 	}
@@ -854,6 +877,7 @@
 .eb-pricebox-price {
 	margin: 0;
 }
+
 .eb-pricebox-feature-button {
 	margin: 0 auto;
 }
@@ -869,6 +893,7 @@
 	display: flex !important;
 	justify-content: center;
 }
+
 .eb-pricebox-sortable-item {
 	display: flex;
 	justify-content: space-between;
@@ -877,6 +902,7 @@
 	border: 1px solid;
 	line-height: 2.5em;
 }
+
 .eb-pricebox-sortable-title {
 	flex: 12;
 	padding-left: 5px;
@@ -885,6 +911,7 @@
 	width: 200px;
 	white-space: nowrap;
 }
+
 .eb-pricebox-sortable-trash {
 	font-size: 16px;
 	border-left: 1px solid gray;
@@ -893,24 +920,30 @@
 	text-align: center;
 	padding: 8px;
 }
+
 .eb-pricebox-sortable-trash:hover {
 	background: #f44336;
 	color: #ffffff;
 	cursor: pointer;
 }
+
 .eb-pricebox-sortable-title:hover {
 	cursor: pointer;
 }
+
 .eb-pricebox-button {
 	line-height: inherit;
 }
+
 .eb-pricebox-add-button-label {
 	margin-left: 10px;
 }
+
 .eb-pricebox-features {
 	margin: 0;
 	padding: 0;
 }
+
 .eb-pricebox-wrapper p {
 	margin-bottom: 0;
 }
@@ -929,7 +962,8 @@
 		transition: 0.5s;
 
 		&:hover {
-			box-shadow: 0 14px 28px rgb(0 0 0 / 25%), 0 10px 10px rgb(0 0 0 / 22%);
+			box-shadow: 0 14px 28px rgb(0 0 0 / 25%),
+				0 10px 10px rgb(0 0 0 / 22%);
 		}
 
 		// ribbon
@@ -975,6 +1009,7 @@
 				text-transform: uppercase;
 				padding: 5px 10px;
 			}
+
 			&:after {
 				content: "";
 				position: absolute;
@@ -1018,6 +1053,7 @@
 				}
 			}
 		}
+
 		&.ribbon-3:before {
 			content: "Featured";
 			position: absolute;
@@ -1032,6 +1068,7 @@
 			text-transform: uppercase;
 			padding: 5px 15px;
 		}
+
 		&.ribbon-3.left:before {
 			left: 15px;
 			right: auto;
@@ -1140,6 +1177,7 @@
 		border-top-left-radius: inherit;
 		border-top-right-radius: inherit;
 	}
+
 	.eb-pricing-bottom {
 		padding: 30px;
 	}
